Specify menu path using script-fu-menu-register
Previously the full path of the menu location of a script-fu entry was
specified as the second parameter of the script-fu-register form. That
behavior is depreciated and only the leaf entry should be specified
here. Instead the menu path (minus the leaf entry) should be specified
in a coordinating script-fu-menu-register form.